idiotDiary_old/requirements.txt
aiogram==3.4.1
aiogram_dialog==2.2.0a3
alembic==1.13.1
annotated-types==0.6.0
APScheduler==3.10.4
asyncpg==0.29.0
babel==2.15.0
environs==11.0.0
eyed3==0.9.7
Jinja2==3.1.3
loguru==0.7.2
Pillow==10.3.0
pytz==2024.1
redis==5.1.0b4
selenium==4.19.0
SQLAlchemy==2.0.29
typing-extensions==4.11.0
yt-dlp==2024.4.9
